refact: step0 2주차 과제 수정중-4

+++ src/main/java/gift/dto/ProductDto.java
@@ -47,7 +47,7 @@ public class ProductDto {
         return result;
     }
 
-    private boolean checkValidLength(String name, Long start, Long end) {
+    private boolean checkValidLength(String name, int start, int end) {
         return (name.length() >= 1 && name.length() <= end);
     }
